refactor: api design 변경
