+++ src/main/java/com/flab/daitso/controller/ProductController.java
@@ -8,6 +8,7 @@ import javax.validation.Valid;
 import java.util.List;
 
 @RestController
+@RequestMapping("/api/products")
 public class ProductController {
 
     private final ProductService productService;
@@ -20,7 +21,7 @@ public class ProductController {
      * 특정 상품을 등록하는 기능
      * @param productDto 상품 정보
      */
-    @PostMapping("/v2/providers/api/v1/seller-products")
+    @PostMapping
     public void registerProduct(@RequestBody @Valid ProductDto productDto) {
         productService.registerProduct(productDto);
     }
@@ -29,7 +30,7 @@ public class ProductController {
      * 하나의 특정 상품을 상품명으로 검색하는 기능
      * @param productId 상품아이디
      */
-    @GetMapping("/v2/providers/api/v1/seller-products/{productId}")
+    @GetMapping("/{productId}")
     public ProductDto searchById(@PathVariable Long productId) {
         return productService.findProductById(productId);
     }
@@ -38,7 +39,7 @@ public class ProductController {
      * 특정 상품을 목록에서 삭제하는 기능
      * @param productId 상품아이디
      */
-    @DeleteMapping("/v2/providers/api/v1/seller-products/{productId}")
+    @DeleteMapping("/{productId}")
     public void deleteProduct(@PathVariable Long productId) {
         productService.deleteProduct(productId);
     }
@@ -49,34 +50,4 @@ public class ProductController {
         productService.saveProductInCategory(categoryId, products);
     }
      */
-
-    @GetMapping("/np/categories/{categoryId}")
-    public List<ProductDto> getProductListBySort(@PathVariable Long categoryId,
-                                                 @RequestParam(value = "score", required = false, defaultValue = "0") Long score,
-                                                 @RequestParam(value = "minPrice", required = false, defaultValue = "0") Long minPrice,
-                                                 @RequestParam(value = "maxPrice", required = false, defaultValue = "0") Long maxPrice,
-                                                 @RequestParam(value = "sorter", required = false, defaultValue = "0") Long sorter) {
-        if (score > 0) {
-            return getProductListByScoreRange(categoryId, score);
-        }
-        if (minPrice > 0 && (minPrice <= maxPrice)) {
-            return getProductListByPriceRange(categoryId, minPrice, maxPrice);
-        }
-        if (sorter >= 0) {
-            return getProductListByLatestOrder(categoryId, sorter);
-        }
-        return productService.findProductListByScoreRange(categoryId, 0L);
-    }
-
-    public List<ProductDto> getProductListByScoreRange(Long categoryId, Long score) {
-        return productService.findProductListByScoreRange(categoryId, score);
-    }
-
-    public List<ProductDto> getProductListByPriceRange(Long categoryId, Long minPrice, Long maxPrice) {
-        return productService.findProductListByPriceRange(categoryId, minPrice, maxPrice);
-    }
-
-    public List<ProductDto> getProductListByLatestOrder(Long categoryId, Long sorter) {
-        return productService.findProductListByLatestOrder(categoryId, sorter);
-    }
 }
