+++ src/test/java/com/flab/daitso/controller/ProductControllerTest.java
@@ -67,7 +67,7 @@ public class ProductControllerTest {
                 .content("test 상품입니다.")
                 .build();
 
-        mvc.perform(post("/v2/providers/api/v1/seller-products")
+        mvc.perform(post("/api/products")
                 .content(objectMapper.writeValueAsString(productDto))
                 .contentType(MediaType.APPLICATION_JSON))
                 .andDo(print())
@@ -86,7 +86,7 @@ public class ProductControllerTest {
 
         ProductDto product = productService.findProductByName("test1");
 
-        mvc.perform(get("/v2/providers/api/v1/seller-products/" + product.getProductId())
+        mvc.perform(get("/api/products/" + product.getProductId())
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andDo(print());
@@ -96,7 +96,7 @@ public class ProductControllerTest {
     @Test
     @DisplayName("상품 아이디 해당 상품을 제대로 검색하는지 테스트")
     public void 존재하지_않는_상품_테스트() throws Exception {
-        mvc.perform(get("/v2/providers/api/v1/seller-products/" + -1)
+        mvc.perform(get("/api/products/" + -1)
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isBadRequest())
                 .andExpect(content().string("NotFoundException"))
@@ -113,63 +113,9 @@ public class ProductControllerTest {
                 .build();
         productService.registerProduct(product);
 
-        mvc.perform(delete("/v2/providers/api/v1/seller-products/" + product.getProductId())
+        mvc.perform(delete("/api/products/" + product.getProductId())
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andDo(print());
-    }
-
-    @Test
-    @DisplayName("별점으로 상품 리스트 반환")
-    public void 별점으로_상품_리스트_반환() throws Exception {
-        List<ProductDto> products1 = new ArrayList<>();
-        List<ProductDto> products2 = new ArrayList<>();
-
-        Category carGoods = new Category("car goods");
-        Category interior = new Category("interior");
-        Category seat = new Category("seat");
-
-        carGoods.addChildCategory(interior);
-        carGoods.addChildCategory(seat);
-
-        Long carGoodsId = categoryService.saveCategory(carGoods);
-        Long interiorId = categoryService.saveCategory(interior);
-        Long seatId = categoryService.saveCategory(seat);
-
-        ProductDto productDto1 = new ProductDto.Builder()
-                .name("test1")
-                .price(10000L)
-                .content("test 상품입니다.")
-                .build();
-
-        ProductDto productDto2 = new ProductDto.Builder()
-                .name("test2")
-                .price(20000L)
-                .content("test2 상품입니다.")
-                .build();
-
-        ProductDto productDto3 = new ProductDto.Builder()
-                .name("test3")
-                .price(30000L)
-                .content("test3 상품입니다.")
-                .build();
-
-        productService.registerProduct(productDto1);
-        productService.registerProduct(productDto2);
-        productService.registerProduct(productDto3);
-
-        products1.add(productDto1);
-        products1.add(productDto2);
-        products2.add(productDto3);
-
-        productService.saveProductInCategory(interiorId, products1);
-        productService.saveProductInCategory(seatId, products2);
-
-        mvc.perform(get("/np/categories/" + interiorId)
-                .param("score", "0")
-                .contentType(MediaType.APPLICATION_JSON))
-                .andDo(print())
-                .andExpect(status().isOk());
-
     }
 }
